削除機能のコンポーネント分けとカスタムフックの追加
